base module (not mine)
This (Run from the Shadow by Darren Grey) is the module i'm going to be
basing my code off of because i've never touched either Lua or a
T-Engine module before. It looks good because it's ASCII and fairly
simple. I did not write any of this code, this is just put here to keep
a record of what code wasn't written by me.